rename: 주말 할인명 변경(specialPolicy -> wekendPolicy)
